+++ InfinityHammer/Hammer.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using HarmonyLib;
 using InfinityTools;
 using ServerDevcommands;
@@ -23,19 +25,20 @@ public static class Hammer
   {
     var build = Helper.GetPlayer().m_buildPieces;
     if (build)
-      build.SetSelected(new(-1, -1));
+      build.m_selectedPiece[(int)build.m_selectedCategory] = new(-1, -1);
   }
   public static void SelectRepair()
   {
     var build = Helper.GetPlayer().m_buildPieces;
     if (build)
-      build.SetSelected(new(0, 0));
+      build.m_selectedPiece[(int)build.m_selectedCategory] = new(0, 0);
   }
   public static void SelectRepairIfEmpty()
   {
     var build = Helper.GetPlayer().m_buildPieces;
-    if (build && build.m_selectedPiece[(int)build.GetSelectedCategory()].x == -1)
-      build.SetSelected(new(0, 0));
+    foreach (var category in build.m_categories)
+      if (build.m_selectedPiece[(int)category].x == -1)
+        build.m_selectedPiece[(int)category] = new(0, 0);
   }
 
   private static bool OriginalUseDurability = false;
@@ -90,20 +93,44 @@ public static class Hammer
   }
   public static void OpenBuildMenu()
   {
-    Equip();
     var player = Helper.GetPlayer();
+    if (EquipInfinityHammer())
+    {
+      // Some mods trigger when place mode is set. So make sure it is retriggered if the hammer is already equipped.
+      player.SetPlaceMode(player.m_buildPieces);
+    }
     var pt = player.m_buildPieces;
     if (pt)
     {
-      pt.m_selectedCategory = 0;
-      pt.m_selectedPiece[0] = new(0, 0);
+      pt.m_selectedCategory = pt.m_categories.Count > 0 ? pt.m_categories[0] : 0;
+      pt.m_selectedPiece[(int)pt.m_selectedCategory] = new(0, 0);
     }
-    player.UpdateAvailablePiecesList();
     Hud.instance.m_pieceSelectionWindow.SetActive(true);
     Hud.instance.m_closePieceSelection = 0;
     Hud.instance.UpdateBuild(Player.m_localPlayer, true);
   }
-
+  public static bool IsInfinityHammer(ItemDrop.ItemData item) => item != null && item.m_customData != null && item.m_customData.ContainsKey("infinity_hammer");
+  public static bool IsInfinityHammer(PieceTable pt) => pt && pt.name == "_InfinityHammerPieceTable";
+  public static bool EquipInfinityHammer()
+  {
+    var player = Helper.GetPlayer();
+    var rightItem = player.GetRightItem();
+    if (rightItem != null && rightItem.m_customData.ContainsKey("infinity_hammer")) return true;
+    var inventory = player.GetInventory();
+    var infinityHammer = inventory.m_inventory.Find(IsInfinityHammer);
+    if (infinityHammer == null)
+    {
+      var freeSlot = inventory.FindEmptySlot(true);
+      var data = new Dictionary<string, string>();
+      data["infinity_hammer"] = "true";
+      if (!inventory.AddItem("Hammer", 1, 100f, freeSlot, false, 1, 0, player.GetPlayerID(), Game.instance.GetPlayerProfile().GetName(), data, 0, true))
+        throw new InvalidOperationException("Unable to add the hammer to inventory.");
+    }
+    infinityHammer = inventory.m_inventory.Find(item => item != null && item.m_customData.ContainsKey("infinity_hammer"));
+    if (infinityHammer == null) throw new InvalidOperationException("Unable to find the hammer in inventory.");
+    player.EquipItem(infinityHammer);
+    return false;
+  }
 
   public static bool Is(ItemDrop.ItemData item) => item != null && item.m_shared.m_buildPieces != null;
 
@@ -117,11 +144,63 @@ public static class Hammer
     var player = Helper.GetPlayer();
     if (!player) return "";
     var item = player.GetRightItem();
+    if (IsInfinityHammer(item)) return "infinity_hammer";
     if (item == null) return "";
     return Utils.GetPrefabName(item.m_dropPrefab).ToLower();
   }
 }
 
+
+[HarmonyPatch]
+public class CustomHammer
+{
+
+  private static Sprite? cachedSprite = null;
+  [HarmonyPatch(typeof(ItemDrop.ItemData), nameof(ItemDrop.ItemData.GetIcon)), HarmonyPostfix]
+  public static Sprite GetIcon(Sprite result, ItemDrop.ItemData __instance)
+  {
+    if (!Hammer.IsInfinityHammer(__instance)) return result;
+    if (cachedSprite == null) cachedSprite = SpriteHelper.FindSprite("+🛠️");
+    return cachedSprite ?? result;
+  }
+
+  [HarmonyPatch(typeof(ItemDrop), nameof(ItemDrop.GetHoverName)), HarmonyPostfix]
+  public static string GetHoverName(string result, ItemDrop __instance)
+  {
+    var item = __instance.m_itemData;
+    return Hammer.IsInfinityHammer(item) ? "Infinity Hammer" : result;
+  }
+
+  private static PieceTable? cachedPieceTable = null;
+
+  [HarmonyPatch(typeof(Player), nameof(Player.SetPlaceMode)), HarmonyPrefix]
+  public static void SetPlaceModePrefix(Player __instance, ref PieceTable buildPieces)
+  {
+    var item = __instance.GetRightItem();
+    if (!Hammer.IsInfinityHammer(item)) return;
+    if (HammerMenuCommand.CurrentMode == MenuMode.Builds && HammerMenuCommand.CurrentFilter != "")
+    {
+      var items = CustomMenu.GetBuildItems();
+      var build = items.FirstOrDefault(i => i.name.Equals(HammerMenuCommand.CurrentFilter, StringComparison.InvariantCultureIgnoreCase));
+      if (build != null)
+      {
+        buildPieces = build.m_itemData.m_shared.m_buildPieces;
+        return;
+      }
+    }
+    if (cachedPieceTable == null)
+    {
+      var obj = new GameObject("_InfinityHammerPieceTable");
+      UnityEngine.Object.DontDestroyOnLoad(obj);
+      var pt = obj.AddComponent<PieceTable>();
+      pt.m_canRemoveFeasts = true;
+      pt.m_canRemovePieces = true;
+      pt.m_skill = Skills.SkillType.None;
+      cachedPieceTable = pt;
+    }
+    buildPieces = cachedPieceTable;
+  }
+}
 
 
 [HarmonyPatch(typeof(Humanoid), nameof(Humanoid.UnequipItem))]