+++ publish/CHANGELOG.md
@@ -1,5 +1,10 @@
+- v1.75
+  - Reverts the fix for error message when placing some item drops (this modified vanilla behavior).
+
 - v1.74
   - Adds a new command "hammer_zoop_reset" to reset zoop offset.
+  - Adds a new command "hammer_mark" to mark individual pieces for selection (experimental).
+  - Adds a new command "hammer_menu" to use a custom UI for selecting pieces (experimental).
   - Adds automatic clean up for nearby corrupted fish.
   - Fixes controller not working on build menu.
   - Fixes trying to place Twig breaking the player entirely.